media-libs/libsdl2: Add vulkan support via USE flag
This now enables the "loadso" option unconditionally, partly because
both Vulkan and Windows require it, and partly because upstream insist
that it should be enabled. It is feasible that Vulkan could be made to
work without it and upstream have shown interest so if someone is
prepared to do the work then we can revisit this.

Closes: https://bugs.gentoo.org/644732
Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11